Disabling all arguments after foreach
Fixes #50.
Uses `DisableFlagParsing` as per the documentation to disable the
subsequent flag parsing in the command and leaving it as arguments to
the shell command.

Added `TraverseChildren` to rootCmd to allow -v and -h to be parsed.

https://pkg.go.dev/github.com/spf13/cobra#Command

Tested on Mac OS.

I'm unable to write tests as we're not testing the flag parsing at the
moment.