Replace emoji with jest-style colors
I've experimented with some true-console experience
and tried to implement some jest-like badges.
That's not so distracting as more or less fits into
terminal color palette and resolution. I would say it
also feels a bit more like a professional tool, but that's for sure IMO only.

Just take a look and please tell me WDYT?