Adds the ability to clone only and not fork (#62)
* Clone does not fork unless specified

With this commit, the `clone` command is defaulting to cloning and
forking only when the user has not push rights onto the repository.

If the `--fork` flag is specified, it will fork and clone
systematically.

* Adding `--no-fork` option and refactoring `clone`.

Adding a `--no-fork` flag and leave the default behaviour as "always
fork". Changing the clone unit test to better represent the two
scenarii.