Adds `update-prs` with `close` feature
Relates to #65
Fixes #60
Relates to #67

This PR implements a global PR closing command.
It introduces a new `update-prs` verb which has only one option
at the moment: `--close`.

It uses the same mechanism as in #67, looking at the PRs and identifying
the one related to the campain, then closes it.

It handles the case where there is no open PR.