Fixing foreach documentation
the `--` were needed for any command including parameters prior to #51.
After #51, the `--` are no longer needed a cobra isn't trying to parse
the full command. It's also failing the run now if the `--` are
included since they are interpreted as part of the command.

This commit fixes the documentation associated with `foreach`.