Fixes the release with goreleaser fixed version
[Replacements](https://github.com/Skyscanner/turbolift/blob/9f28424daf82663b91716b6888c12cfb6a04ca0e/.goreleaser.yml#L16-L21) have been deprecated on goreleaser since this [commit](https://github.com/goreleaser/goreleaser/pull/3589/files#diff-5563ff2aab3eb7eae7bd8557cd2c086d8373263762684283485b2f0a00a0f52a).
The github action for goreleaser unfortunately uses the latest
goreleaser version instead of a fixed version.

We need to address the replacements in a future commit, for now we are just forcing an older version of goreleaser.
The version of goreleaser that includes the deprecation of `replacements` is [v1.19.0](https://github.com/goreleaser/goreleaser/releases/tag/v1.19.0).
We are rolling back now to [v1.18.2](https://github.com/goreleaser/goreleaser/releases/tag/v1.18.2).