fix(executor): Adding silent mode and tests
While looking at #94, I've notived that the output of `turbolift commit`
includes a weird line:

```
Executing: /opt/homebrew/bin/zsh [-c ...]
```

This is because `isRepoChanged` function executes a command to figure out whether something is marked as changed. Every command executed generates an echo of the command for the user.

This PR introduces a new flag on the executor that will skip the echo.
This fixes the extraneous output in `isRepoChanged` but will also be
useful in future features.