feat(GHA): adding conventional commit linting
Adding a new github action step as linting for commit messages.
Lots of OS repositories seem to have adopted the [conventional commit
style] (https://www.conventionalcommits.org/en/v1.0.0/#summary).

This PR is a simple check on the commit message, giving the commiter
feedback on the commit message style.

This could be complemented by a precommit check if we revisit it.